predict/pyproject.toml
[project]
name = "predictclaw"
version = "0.1.0"
description = "PredictClaw isolated predict.fun skill package"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "predict-sdk>=0.0.15",
    "eth-account>=0.13.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.11.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0.2",
    "respx>=0.22.0",
    "web3>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:websockets\\.legacy is deprecated:DeprecationWarning",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
reportMissingModuleSource = false
